[Examples] Create common tv-app sources
Move cluster handlers, app platform and shell command implementations,
ZCL callbacks from Linux to tv-common directory.
Create source set with tv-app source files.
Create cmake macro to add common tv-app sources to the
specific target.
Use common tv-app sources in Linux tv-app example.

Sources can now be easily shared across platform implementations.
