+++ examples/tv-app/tv-common/clusters/application-launcher/ApplicationLauncherManager.cpp
