+++ examples/tv-app/tv-common/clusters/content-launcher/ContentLauncherManager.h
@@ -59,7 +59,6 @@ protected:
     std::vector<ContentEntry> mContentList;
 
 private:
-    EndpointId mEndpointId;
     // TODO: set this based upon meta data from app
     uint32_t mDynamicEndpointFeatureMap = 3;
 };
