+++ examples/tv-app/tv-common/include/CHIPProjectAppConfig.h
@@ -70,8 +70,5 @@
 // Change port to make it easy to run against tv-casting-app
 #define CHIP_PORT 5640
 
-// Create a dedicated file for storage to make it easy to run against other apps
-// #define CHIP_CONFIG_KVS_PATH "/tmp/chip_tv_kvs"
-
-// include the CHIPProjectConfig from config/standalone
+// Include the CHIPProjectConfig from platform implementation config
 #include <CHIPProjectConfig.h>
